+++ app/api/v2/Keyvalue.php
@@ -8,12 +8,20 @@
 
 namespace app\api\v2;
 
-use \app\inc\Route;
-use \app\inc\Input;
+use app\inc\Controller;
+use app\inc\Route;
+use app\inc\Input;
+use app\inc\Util;
 
-
-class Keyvalue extends \app\inc\Controller
+/**
+ * Class Keyvalue
+ * @package app\api\v2
+ */
+class Keyvalue extends Controller
 {
+    /**
+     * @var \app\models\Keyvalue
+     */
     public $keyValue;
 
     function __construct()
@@ -23,12 +31,12 @@ class Keyvalue extends \app\inc\Controller
     }
 
     /**
-     * @return array
+     * @return array<mixed>
      *
      * @OA\Get(
-     *   path="/api/v2/keyvalue/{userId}",
+     *   path="/api/v2/keyvalue/{userId}/{key}",
      *   tags={"Keyvalue"},
-     *   summary="Coming",
+     *   summary="Get value by key",
      *   @OA\Parameter(
      *     name="userId",
      *     in="path",
@@ -38,40 +46,238 @@ class Keyvalue extends \app\inc\Controller
      *       type="string"
      *     )
      *   ),
+     *   @OA\Parameter(
+     *     name="key",
+     *     in="path",
+     *     required=false,
+     *     description="Key to fetch",
+     *     @OA\Schema(
+     *       type="string"
+     *     )
+     *   ),
      *   @OA\Response(
      *     response="200",
-     *     description="Operation status"
+     *     description="Operation status",
+     *     @OA\MediaType(
+     *       mediaType="application/json",
+     *       @OA\Schema(
+     *         type="object",
+     *         @OA\Property(property="success", type="boolean", example=true),
+     *         @OA\Property(property="data", type="object")
+     *       )
+     *     ),
+     *     @OA\MediaType(
+     *       mediaType="text/plain",
+     *       @OA\Schema(
+     *         type="string",
+     *       )
+     *     )
      *   )
      * )
      */
     public function get_index(): array
     {
         $key = Route::getParam("key");
-        return $this->keyValue->get($key,  Input::get());
+        $data = $this->keyValue->get($key, Input::get());
+        if (Input::getAccept() == Input::TEXT_PLAIN) {
+            $data = Util::base64urlEncode(json_encode($data));
+            return [
+                "text" => $data
+            ];
+        }
+        return $data;
     }
 
     /**
-     * @return array
+     * @return array<mixed>
+     *
+     * @OA\Post(
+     *   path="/api/v2/keyvalue/{userId}/{key}",
+     *   tags={"Keyvalue"},
+     *   summary="Create new key/value in store",
+     *   @OA\Parameter(
+     *     name="userId",
+     *     in="path",
+     *     required=true,
+     *     description="User identifier",
+     *     @OA\Schema(
+     *       type="string"
+     *     )
+     *   ),
+     *   @OA\Parameter(
+     *     name="key",
+     *     in="path",
+     *     required=true,
+     *     description="Key to create",
+     *     @OA\Schema(
+     *       type="string"
+     *     )
+     *   ),
+     *   @OA\RequestBody(
+     *     description="JSON value to store",
+     *     @OA\MediaType(
+     *       mediaType="application/json",
+     *       @OA\Schema(
+     *         type="object"
+     *       )
+     *     ),
+     *     @OA\MediaType(
+     *       mediaType="text/plain",
+     *       @OA\Schema(
+     *         type="string"
+     *       )
+     *     )
+     *   ),
+     *   @OA\Response(
+     *     response="200",
+     *     description="Operation status",
+     *     @OA\MediaType(
+     *       mediaType="application/json",
+     *       @OA\Schema(
+     *         type="object",
+     *         @OA\Property(property="success", type="boolean", example=true),
+     *         @OA\Property(property="data", type="object")
+     *       )
+     *     ),
+     *     @OA\MediaType(
+     *       mediaType="text/plain",
+     *       @OA\Schema(
+     *         type="string",
+     *       )
+     *     )
+     *   )
+     * )
      */
     public function post_index(): array
     {
         $key = Route::getParam("key");
         $json = Input::getBody();
-        return $this->keyValue->insert($key, $json);
+        if (Input::getContentType() == Input::TEXT_PLAIN) {
+            $json = Util::base64urlDecode($json);
+        }
+        $data = $this->keyValue->insert($key, $json);
+        if (Input::getAccept() == Input::TEXT_PLAIN) {
+            return [
+                "text" => Util::base64urlEncode(json_encode($data))
+            ];
+        }
+        return $data;
     }
 
     /**
-     * @return array
+     * @return array<mixed>
+     *
+     * @OA\Put(
+     *   path="/api/v2/keyvalue/{userId}/{key}",
+     *   tags={"Keyvalue"},
+     *   summary="Update value in store",
+     *   @OA\Parameter(
+     *     name="userId",
+     *     in="path",
+     *     required=true,
+     *     description="User identifier",
+     *     @OA\Schema(
+     *       type="string"
+     *     )
+     *   ),
+     *   @OA\Parameter(
+     *     name="key",
+     *     in="path",
+     *     required=true,
+     *     description="Key to update",
+     *     @OA\Schema(
+     *       type="string"
+     *     )
+     *   ),
+     *   @OA\RequestBody(
+     *     description="New JSON value to store",
+     *     @OA\MediaType(
+     *       mediaType="application/json",
+     *       @OA\Schema(
+     *         type="object"
+     *       )
+     *     ),
+     *     @OA\MediaType(
+     *       mediaType="text/plain",
+     *       @OA\Schema(
+     *         type="string"
+     *       )
+     *     )
+     *   ),
+     *   @OA\Response(
+     *     response="200",
+     *     description="Operation status",
+     *     @OA\MediaType(
+     *       mediaType="application/json",
+     *       @OA\Schema(
+     *         type="object",
+     *         @OA\Property(property="success", type="boolean", example=true),
+     *         @OA\Property(property="data", type="object")
+     *       )
+     *     ),
+     *     @OA\MediaType(
+     *       mediaType="text/plain",
+     *       @OA\Schema(
+     *         type="string",
+     *       )
+     *     )
+     *   )
+     * )
      */
     public function put_index(): array
     {
         $key = Route::getParam("key");
         $json = Input::getBody();
-        return $this->keyValue->update($key, $json);
+        if (Input::getContentType() == Input::TEXT_PLAIN) {
+            $json = Util::base64urlDecode($json);
+        }
+        $data = $this->keyValue->update($key, $json);
+        if (Input::getAccept() == Input::TEXT_PLAIN) {
+            return [
+                "text" => Util::base64urlEncode(json_encode($data))
+            ];
+        }
+        return $data;
     }
 
     /**
-     * @return array
+     * @return array<mixed>
+     *
+     * @OA\Delete(
+     *   path="/api/v2/keyvalue/{userId}/{key}",
+     *   tags={"Keyvalue"},
+     *   summary="Delete a key/bvalue in store",
+     *   @OA\Parameter(
+     *     name="userId",
+     *     in="path",
+     *     required=true,
+     *     description="User identifier",
+     *     @OA\Schema(
+     *       type="string"
+     *     )
+     *   ),
+     *   @OA\Parameter(
+     *     name="key",
+     *     in="path",
+     *     required=true,
+     *     description="Key to delete",
+     *     @OA\Schema(
+     *       type="string"
+     *     )
+     *   ),
+     *   @OA\Response(
+     *     response="200",
+     *     description="Operation status",
+     *     @OA\MediaType(
+     *       mediaType="application/json",
+     *       @OA\Schema(
+     *         type="object",
+     *         @OA\Property(property="data", type="string", example="my_key"),
+     *         @OA\Property(property="success", type="boolean", example=true)
+     *       )
+     *     )
+     *   )
+     * )
      */
     public function delete_index(): array
     {